Added functionality to create a new db

+++ powershell_scripts/db_cloning_automation/clone_db.ps1
@@ -5,7 +5,9 @@ param(
 [Parameter(Mandatory=$False)]
 [string]$sdb,
 [Parameter(Mandatory=$False)]
-[switch]$backup
+[switch]$backup,
+[Parameter(Mandatory=$False)]
+[switch]$new
 )
 
 # Getting database information from json file
@@ -17,38 +19,44 @@ $allDbsJson = ConvertFrom-Json "$(get-content $dbFilePath)"
 #Login-AzureRmAccount
 Select-AzureRmSubscription -SubscriptionID $allDbsJson.subscriptionId
 
-# Storage account info to store BACPAC
-$BaseStorageUri = "https://ancloudops.blob.core.windows.net/dbbackups/"
-$StorageKeytype = "StorageAccessKey"
-$StorageKey = "VHt9S600ZjY4qfUb0iWyC6Rswf6uVDSsoIIicY4n5GDyiIjJ/rE4GRV9meAZvk9L97IqzNfQr2m1u51beCP2wQ=="
-$bacpacFilename = $allDbsJson.Databases.$ddb.dbname + (Get-Date).ToString("yyyy-MM-dd-HH-mm") + ".bacpac"
-$BacpacUri = $BaseStorageUri + $bacpacFilename
-
-# exporting db to bacpac
-$exportRequest = New-AzureRmSqlDatabaseExport -ResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup $allDbsJson.Databases.$ddb.server `
-                    -DatabaseName $allDbsJson.Databases.$ddb.dbname -StorageKeytype $StorageKeytype -StorageKey $StorageKey -StorageUri $BacpacUri `
-                    -AdministratorLogin $allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).username `
-                    -AdministratorLoginPassword ($allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).password | ConvertTo-SecureString -AsPlainText -Force )
-
-Write-Host "Backup of" $allDbsJson.Databases.$ddb.dbname "is in progress" 
-$exportStatus = Get-AzureRmSqlDatabaseImportExportStatus -OperationStatusLink $exportRequest.OperationStatusLink
-while ($exportStatus.Status -ne 'Succeeded')
+if(-not($new))
 {
-    start-sleep -s 5
+    # Storage account info to store BACPAC
+    $BaseStorageUri = "https://ancloudops.blob.core.windows.net/dbbackups/"
+    $StorageKeytype = "StorageAccessKey"
+    $StorageKey = "VHt9S600ZjY4qfUb0iWyC6Rswf6uVDSsoIIicY4n5GDyiIjJ/rE4GRV9meAZvk9L97IqzNfQr2m1u51beCP2wQ=="
+    $bacpacFilename = $allDbsJson.Databases.$ddb.dbname + (Get-Date).ToString("yyyy-MM-dd-HH-mm") + ".bacpac"
+    $BacpacUri = $BaseStorageUri + $bacpacFilename
+
+
+    # exporting db to bacpac
+    $exportRequest = New-AzureRmSqlDatabaseExport -ResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup $allDbsJson.Databases.$ddb.server `
+                        -DatabaseName $allDbsJson.Databases.$ddb.dbname -StorageKeytype $StorageKeytype -StorageKey $StorageKey -StorageUri $BacpacUri `
+                        -AdministratorLogin $allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).username `
+                        -AdministratorLoginPassword ($allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).password | ConvertTo-SecureString -AsPlainText -Force )
+
+    Write-Host "Backup of" $allDbsJson.Databases.$ddb.dbname "is in progress" 
     $exportStatus = Get-AzureRmSqlDatabaseImportExportStatus -OperationStatusLink $exportRequest.OperationStatusLink
-    Write-Host "Backup of " $allDbsJson.Databases.$ddb.dbname " is in progress" 
+    while ($exportStatus.Status -ne 'Succeeded')
+    {
+        start-sleep -s 5
+        $exportStatus = Get-AzureRmSqlDatabaseImportExportStatus -OperationStatusLink $exportRequest.OperationStatusLink
+        Write-Host "Backup of " $allDbsJson.Databases.$ddb.dbname " is in progress" 
+    }
+
+    Write-Host "Backup completed, URL:" $BacpacUri -ForegroundColor Green
+
 }
-
-Write-Host "Backup completed, URL:" $BacpacUri -ForegroundColor Green
-
 
 
 if(-not($backup))
 {
-    Write-Host "Deleting database" $allDbsJson.Databases.$ddb.dbname -ForegroundColor Yellow
-    Remove-AzureRmSqlDatabase -ResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup -ServerName $allDbsJson.Databases.$ddb.server `
-        -DatabaseName $allDbsJson.Databases.$ddb.dbname -Force
-
+    if(-not($new))
+    {
+        Write-Host "Deleting database" $allDbsJson.Databases.$ddb.dbname -ForegroundColor Yellow
+        Remove-AzureRmSqlDatabase -ResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup -ServerName $allDbsJson.Databases.$ddb.server `
+            -DatabaseName $allDbsJson.Databases.$ddb.dbname -Force
+    }
     Write-Host "Copying database to destination" -ForegroundColor Yellow
 
     # Copy source database to the target server
@@ -57,6 +65,7 @@ if(-not($backup))
                         -CopyServerName $allDbsJson.Databases.$ddb.server -CopyDatabaseName $allDbsJson.Databases.$ddb.dbname `
                         -ServiceObjectiveName "S0"
 
+    start-sleep -s 5
     # Adding to the respective elastic pool
     Write-Host "Adding database to Elastic Pool if any" -ForegroundColor Yellow
     if ($allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).ElasticPoolName -ne "NA")
